% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{ancestral_edges}
\alias{ancestral_edges}
\title{For a node in a tree, get a vector of the edges between the 
node and the root}
\usage{
ancestral_edges(phy, node)
}
\arguments{
\item{phy}{A phylo object}

\item{node}{A node number}
}
\value{
A vector of edge numbers
}
\description{
For a node in a tree, get a vector of the edges between the 
node and the root
}
