% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{are_bipartitions_compatible}
\alias{are_bipartitions_compatible}
\title{Check if two bipartitions drawn from trees with the same tips are compatible with eachother.
Each bipartition is defined as a vector of the names of the tips on one side of the
bipartition.}
\usage{
are_bipartitions_compatible(bi1, bi2, phy)
}
\arguments{
\item{bi1}{The first bipartition.}

\item{bi2}{The second bipartition.}

\item{phy}{A phylo object describing a tree that includes all tips under investigation.
This is used to infer the other half of each bipartition.}
}
\value{
TRUE if bi1 is compatible with bi2, otherwise FALSE.
}
\description{
Check if two bipartitions drawn from trees with the same tips are compatible with eachother.
Each bipartition is defined as a vector of the names of the tips on one side of the
bipartition.
}
