% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{bipartition_for_edge}
\alias{bipartition_for_edge}
\title{Get a bipartition, described as a vector of tip numbers, from a specified tree and
edge number.}
\usage{
bipartition_for_edge(phy, edge)
}
\arguments{
\item{phy}{A phylo object that specifies the tree.}

\item{edge}{The number of the edge that defines the bipartition.}
}
\value{
A vector of tip nodes (specified by numbers) that define one half of the
bipartition (the other half is the set of tip nodes that are not in this vector).
}
\description{
Get a bipartition, described as a vector of tip numbers, from a specified tree and
edge number.
}
