% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{calc_diffs}
\alias{calc_diffs}
\title{For each internal node, calculate the difference in state values
at the two child nodes}
\usage{
calc_diffs(phy, states)
}
\arguments{
\item{phy}{An ape::phylo object}

\item{states}{A vector with length equal to nodes with the state at each node}
}
\value{
A vector of values corresponding to each internal node with the value difference
}
\description{
For each internal node, calculate the difference in state values
at the two child nodes
}
