% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{compatible_edges}
\alias{compatible_edges}
\title{Identify the edges in one phylo object that are compatible with the edges in another
phylo object. Requires the same tip labels for each tree.}
\usage{
compatible_edges(phy1, phy2)
}
\arguments{
\item{phy1}{The tree under consideration}

\item{phy2}{The tree to be compared to}
}
\value{
A boolean vector corresponding to the edges in phy1. Each element is FALSE if
the edge is iscompatible with phy2, or TRUE if compatible.
}
\description{
Identify the edges in one phylo object that are compatible with the edges in another
phylo object. Requires the same tip labels for each tree.
}
