% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{connecting_edges}
\alias{connecting_edges}
\title{For two nodes in a tree, get a vector of the edges that
connect them}
\usage{
connecting_edges(phy, node_a, node_b)
}
\arguments{
\item{phy}{A phylo object}

\item{node_a}{Number of the first node}

\item{node_b}{Number of the second node}
}
\value{
A vector of edge numbers
}
\description{
For two nodes in a tree, get a vector of the edges that
connect them
}
