% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{cut_tree}
\alias{cut_tree}
\title{Cuts a single tree on the branch subtending a specified node}
\usage{
cut_tree(phy, x)
}
\arguments{
\item{phy}{The tree to be cut, as an ape phylo object}

\item{x}{An internal node number. The tree phy will be cut on the branch 
that subtends this nodes.}
}
\value{
A list of phylo objects that are the subtrees
}
\description{
Cuts a single tree on the branch subtending a specified node
}

