% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{decompose_tree}
\alias{decompose_tree}
\title{Decomposes a single tree into a series of subtrees designated by internal
node numbers}
\usage{
decompose_tree(phy, x)
}
\arguments{
\item{phy}{The tree to be decomposed, as an ape phylo object}

\item{x}{A vector of internal node numbers. The tree phy will be cut on each branch
that subtends each of these nodes.}
}
\value{
A list of phylo objects
}
\description{
Decomposes a single tree into a series of subtrees designated by internal
node numbers
}
