% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{descendants}
\alias{descendants}
\title{Get all the descendants of a given node in a tree.}
\usage{
descendants(phy, a, keep_node = FALSE)
}
\arguments{
\item{phy}{A phylo object that specifies the tree.}

\item{a}{The number of a node in phy.}

\item{keep_node}{If FALSE, do not include a in the result.}
}
\value{
A vector of nodes (specified by number) that are descendants of a. Includes
internal and tip nodes.
}
\description{
Get all the descendants of a given node in a tree.
}

