% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{difference_from_calibrated}
\alias{difference_from_calibrated}
\title{Assesses how much phy deviates from an ultrametric tree}
\usage{
difference_from_calibrated(phy, model = "discrete", ...)
}
\arguments{
\item{phy}{A phylo object}

\item{model}{The model used for fitting. "discrete" is used by default for speed}

\item{...}{Additional chronos arguments}
}
\value{
The sum of absolute changes in branch length required
to make an ape::chronos time calbrated tree, normalized by the total branch
length of the calibrated tree. The higher the value, the more the
tree deviates from the calibrated tree.
}
\description{
Assesses how much phy deviates from an ultrametric tree
}
