% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{extend_terminal_branches}
\alias{extend_terminal_branches}
\title{Extends each terminal branch by specified length}
\usage{
extend_terminal_branches(phy, x)
}
\arguments{
\item{phy}{A phylogeny in ape::phylo format}

\item{x}{Amount to extend each branch by}
}
\value{
A phylogeny in ape::phylo format
}
\description{
Extends each terminal branch by specified length
}
