% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{flip_bipartition}
\alias{flip_bipartition}
\title{Given a tree and a bipartition, described as a vector of tip labels on one side of
of the bipartition, return the same bipartition but defined by the tip labels on
the other side of the bipartition.}
\usage{
flip_bipartition(phy, bi)
}
\arguments{
\item{phy}{A phylo object that specifies the tree.}

\item{bi}{The bipartition.}
}
\value{
A vector of tip nodes (specified by labels) that define one half of the
bipartition (the other half is the set of tip nodes that are provided as bi).
}
\description{
Given a tree and a bipartition, described as a vector of tip labels on one side of
of the bipartition, return the same bipartition but defined by the tip labels on
the other side of the bipartition.
}
