% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{generate_constaint_tree}
\alias{generate_constaint_tree}
\title{Generates a tree with a single resolved bipartition between two sets of
tip names. Useful for generating constraint trees.}
\usage{
generate_constaint_tree(tips1, tips2)
}
\arguments{
\item{tips1}{A vector of tip names for clade 1}

\item{tips2}{A vector of tip names for clade 2}
}
\value{
A phylo object
}
\description{
Generates a tree with a single resolved bipartition between two sets of
tip names. Useful for generating constraint trees.
}
\examples{
library( ape )
tips1 = c("a", "b", "c")
tips2 = c("d", "e", "f")
ctree = generate_constaint_tree( tips1, tips2 )
}
