% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{get_bipartitions}
\alias{get_bipartitions}
\title{Get a list of all the bipartitions in a tree.}
\usage{
get_bipartitions(phy)
}
\arguments{
\item{phy}{A phylo object that specifies the tree.}
}
\value{
A list of bipartitions for the tree. The order of the list corresponds to the
edges in phy$edge. Bipartitions are specified as a vector of the tip labels that make
up one half of the bipartition.
}
\description{
Get a list of all the bipartitions in a tree.
}

