% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{get_corresponding_nodes}
\alias{get_corresponding_nodes}
\title{Given two trees phy1 and phy2 with the same topology and tip labels, 
get a vector that indicates which node numbers in phy2 correspond to 
the nodes in phy1}
\usage{
get_corresponding_nodes(phy1, phy2)
}
\arguments{
\item{phy1}{A phylo object}

\item{phy2}{A phylo object}
}
\value{
A numeric vector in the order of nodes in phy1, providing 
corresponding node numbers from phy2
}
\description{
Given two trees phy1 and phy2 with the same topology and tip labels, 
get a vector that indicates which node numbers in phy2 correspond to 
the nodes in phy1
}
