% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hutan.R
\docType{package}
\name{hutan}
\alias{hutan}
\alias{hutan-package}
\title{hutan: A collection of tools for phylogenetic tree manipulation.}
\description{
The hutan package provides functions for common phylogenetic tree 
manipulation tasks, and uses these fascilitate some more specialized 
tasks. It is named after the Indonesian word for forest.
}

