% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{is_monophyletic}
\alias{is_monophyletic}
\title{Test if a set of tips, specified as a vector of node numbers, forms a monophyletic
group in a given tree. The test is unrooted, i.e. the group can span the root.}
\usage{
is_monophyletic(phy, x)
}
\arguments{
\item{phy}{The tree under consideration}

\item{x}{A vector of the node numbers of the tips in question}
}
\value{
A boolean, TRUE if the tips form a monophyletic group.
}
\description{
Test if a set of tips, specified as a vector of node numbers, forms a monophyletic
group in a given tree. The test is unrooted, i.e. the group can span the root.
}

