% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{map_frequency_to_subtree}
\alias{map_frequency_to_subtree}
\title{Map the split frequency from a set of trees with more than 
one sequence per taxon to the edge labels of a tree with one 
tip per taxon}
\usage{
map_frequency_to_subtree(tree_set, focal_tree)
}
\arguments{
\item{tree_set}{A list of phylo object, with one or more tips per
taxon}

\item{focal_tree}{A phylo object, with one tip per taxon}
}
\value{
A phylo object with same topology as focal_tree 
and edge labels that indicate the frequency of each edge 
including tip edges
}
\description{
Map the split frequency from a set of trees with more than 
one sequence per taxon to the edge labels of a tree with one 
tip per taxon
}
