% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{safe.drop.tip}
\alias{safe.drop.tip}
\title{Drops specified tips from a phylogeny. Like ape's drop.tip(), but it works when only a single tip is
to be retained.}
\usage{
safe.drop.tip(phy, tip)
}
\arguments{
\item{phy}{The tree, as an ape phylo object}

\item{x}{A vector of tip numbers to be removed.}
}
\value{
The reduced tree, as a phylo object
}
\description{
Drops specified tips from a phylogeny. Like ape's drop.tip(), but it works when only a single tip is
to be retained.
}

