% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{sim_diffs}
\alias{sim_diffs}
\title{For each internal node, simulate the difference between state
values at the node's children}
\usage{
sim_diffs(phy, model_parameters, model_method = "BM")
}
\arguments{
\item{phy}{An ape::phylo object}

\item{model_parameters}{Parameter estimates for evolutionary model}

\item{model_method}{The model of trait evolution. Can be one of c("BM","OU")}
}
\value{
The simulated child differences for each internal node
}
\description{
For each internal node, simulate the difference between state
values at the node's children
}
