% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/siphonophore_constraint.R
\docType{data}
\name{siphonophore_constraint}
\alias{siphonophore_constraint}
\title{Siphonophores constraint phylogeny.}
\format{
An ape phylo object
}
\source{
\url{http://dx.doi.org/10.1080/10635150500354837}
}
\usage{
siphonophore_constraint
}
\description{
An unresolved phylogeny that constrains the group 
Agalmatidae sensu stricto + Bargmannia to be monophyletic, corresponding to 
the published SOWH test
}
\keyword{datasets}
