% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/siphonophore_ml.R
\docType{data}
\name{siphonophore_ml}
\alias{siphonophore_ml}
\title{Siphonophores phylogeny.}
\format{An ape phylo object}
\source{
\url{http://dx.doi.org/10.1080/10635150500354837}
}
\usage{
siphonophore_ml
}
\description{
A maximum likelihood phylogeny of siphonophores
}
\keyword{datasets}

