% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{tip_descendants}
\alias{tip_descendants}
\title{Get all the tips that are descendants of a given node in a tree.}
\usage{
tip_descendants(phy, a)
}
\arguments{
\item{phy}{A phylo object that specifies the tree.}

\item{a}{The number of a node in phy.}
}
\value{
A vector of tip nodes (specified by number) that are descendants of a. If a is
a tip, it is the sole element of this vector.
}
\description{
Get all the tips that are descendants of a given node in a tree.
}

