% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{tips}
\alias{tips}
\title{Get tips and labels of a phylo object.}
\usage{
tips(phy)
}
\arguments{
\item{phy}{A phylo object.}
}
\value{
A vector of all the tips, annotated with their names
}
\description{
Get tips and labels of a phylo object.
}
