% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{zero_constrained}
\alias{zero_constrained}
\title{Generates the "zero-constrained" tree described by Susko 2014 
(http://dx.doi.org/10.1093/molbev/msu039)}
\usage{
zero_constrained(phy_resolved, phy_constraint, epsilon = 1e-06)
}
\arguments{
\item{phy_resolved}{A fully resolved phylogeny stored as a phylo object, e.g. an ML 
tree.}

\item{phy_constraint}{A partially resolved constraint tree.}

\item{epsilon}{The value to replace the branch length with}
}
\value{
A phylo object containing a tree that is the same as phy_resolved, except that 
the length of edges that are incompatible with phy_constraint are replaced with epsilon.
}
\description{
Generates the "zero-constrained" tree described by Susko 2014 
(http://dx.doi.org/10.1093/molbev/msu039)
}
\examples{
data( siphonophore_ml )
data( siphonophore_constraint )
zc <- zero_constrained( siphonophore_ml, siphonophore_constraint )
}

