% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{picx}
\alias{picx}
\title{Estimate the extended phylogenetic independent contrast. Rather 
than normalize differences across nodes by branch lengths, differences
are normalized be the expected difference obtained from replicate 
simulations. This allows for greater model flexibility.}
\usage{
picx(x, phy, var.contrasts = FALSE, model_method = "BM",
  model_parameters = NA, n_replicates = 50)
}
\arguments{
\item{x}{A numeric vector with one trait value per tip}

\item{phy}{An ape::phylo object}

\item{var.contrasts}{logical, indicates whether the expected variances of the contrasts should be returned}

\item{model_method}{The model of trait evolution. Can be one of c("BM", "OU")}

\item{model_parameters}{Model parameters from fitContinuous. Will estimate if not provided.}

\item{n_replicates}{The number of simulations used to estimate the expected differences}
}
\value{
A vector of phylogenetically independent contrasts
}
\description{
Estimate the extended phylogenetic independent contrast. Rather 
than normalize differences across nodes by branch lengths, differences
are normalized be the expected difference obtained from replicate 
simulations. This allows for greater model flexibility.
}
